move authRoute logic to autController
