+++ client/vite.config.js
@@ -13,4 +13,5 @@ export default defineConfig({
       },
     },
   },
+  build: { chunkSizeWarningLimit: 4000, },
 });
