--- server/controller/userController.js
@@ -1,31 +1,0 @@
-const User = require("../models/User");
-
-const bcrypt = require("bcryptjs");
-const jwt = require("jsonwebtoken");
-
-const authController = async (req, res) => {
-  try {
-    const user = await User.findById({ _id: req.body.userId });
-    user.password = undefined;
-    if (!user) {
-      return res.status(200).send({
-        message: "user not found",
-        success: false,
-      });
-    } else {
-      res.status(200).send({
-        success: true,
-        data: user,
-      });
-    }
-  } catch (error) {
-    console.log(error);
-    res.status(500).send({
-      message: "auth error",
-      success: false,
-      error,
-    });
-  }
-};
-
-module.exports = { authController };
